requirements.txt
requests==2.25.1
nltk==3.6.2
jsonlines==2.0.0
grpcio==1.37.1
grpcio-tools==1.37.1
penman==1.2.0
fuzzywuzzy==0.17.0
networkx==2.4
spacy==2.0.12
neuralcoref==3.1
gensim==3.8.1
protobuf==3.17.3

numpy
python-Levenshtein
flask
matplotlib
graphviz
ipykernel
SPARQLWrapper
ipdb
pyyaml
notebook
autoflake
isort
prettytable

# third party dependencies
torch==1.3
# fairseq==0.8.0
packaging
